refactor: fix resync base models
