+++ cognite/powerops/resync/models/diff/_changes.py
@@ -13,7 +13,7 @@ from cognite.client.data_classes._base import CogniteResource
 from deepdiff import DeepDiff  # type: ignore[import]
 
 from cognite.powerops.resync.models.base.resource_type import Resource
-from cognite.powerops.resync.models.cdf_resources import CDFResource
+from cognite.powerops.resync.models.base.cdf_resources import CDFResource
 
 
 @dataclass
